requirements.txt
streamlit
requests
openai
pandas
openpyxl
anthropic
google-genai
mistralai
supabase
PyYAML
textstat
numpy
plotly
